--- a/middleware/queue/include/queue/group/queuebase/statement.h
+++ b/middleware/queue/include/queue/group/queuebase/statement.h
@@ -37,6 +37,8 @@ namespace casual
          sql::database::Statement commit3;
 
          sql::database::Statement rollback1;
+
+         //! arguments: gtrid, now
          sql::database::Statement rollback2;
          sql::database::Statement rollback3;