Before: we check that more than 100ms has passed ` > 100ms`

Now: we chack that at least 100ms has passed ` >= 100ms`

On some machines it takes less than 1ms (.5ms) to rollback and dequeue the message.

Resolves #497